+++ packages/insomnia-inso/package.json
@@ -65,7 +65,6 @@
     "insomnia-plugin-core-themes": "^3.19.0-beta.4",
     "insomnia-plugin-default-headers": "^3.19.0-beta.4",
     "insomnia-plugin-request": "^3.19.0-beta.4",
-    "insomnia-plugin-response": "^3.19.0-beta.4",
     "insomnia-send-request": "^3.19.0-beta.4",
     "insomnia-testing": "^3.19.0-beta.4",
     "nedb": "^1.8.0",
