+++ packages/insomnia/src/templating/extensions/index.ts
@@ -1,3 +1,4 @@
+import { ExtraRenderInfo } from '../../common/render';
 import type { Request } from '../../models/request';
 import type { Response } from '../../models/response';
 import {
@@ -68,6 +69,9 @@ export type PluginArgument =
 export type PluginTemplateTagContext = HelperContext & {
   app: AppContext;
   store:  PluginStore;
+  network: {
+    sendRequest(request: Request, extraInfo?: ExtraRenderInfo): Promise<Response>;
+  };
   util: {
     models: {
       request: {
