--- plugins/insomnia-plugin-response/README.md
@@ -1,3 +1,0 @@
-# Insomnia Response Template Tag
-
-This is a core Insomnia plugin.
