--- plugins/insomnia-plugin-response/package.json
@@ -1,30 +1,0 @@
-{
-  "private": true,
-  "name": "insomnia-plugin-response",
-  "version": "3.19.0-beta.4",
-  "author": "Kong <office@konghq.com>",
-  "description": "Insomnia response template tag",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Kong/insomnia.git",
-    "directory": "plugins/insomnia-plugin-response"
-  },
-  "bugs": {
-    "url": "https://github.com/Kong/insomnia/issues"
-  },
-  "main": "index.js",
-  "insomnia": {
-    "name": "response",
-    "unlisted": true
-  },
-  "scripts": {
-    "test": "jest --silent"
-  },
-  "dependencies": {
-    "iconv-lite": "^0.6.3",
-    "jsonpath-plus": "^6.0.1",
-    "xmldom": "^0.6.0",
-    "xpath": "0.0.32"
-  }
-}
